### 2020-03-01 #2132 @thesketh
From a quick glance at the list of reviewers, RMeli and amandadumi seem to potentially be good reviewers as both have experience in computational chemistry. The DOI suggested for the VMD article seems correct, I'll make sure that I update it!
