### 2020-03-02 thesketh/pygen-structures#2 @thesketh
Documentation needs more examples
At present, the ReadTheDocs documentation is somewhat limited. It contains the function/class signatures, but no examples of using the code.

There _are_ examples in the README which cover the most basic use, but these could be expanded upon for more complex examples.
