### 2020-03-02 thesketh/pygen-structures#3 @thesketh
Unit and integration tests are lacking
At present, the unit tests mostly cover the `charmm_containers` module. This is intentional - these classes have been the source of most of the issues I've discovered in the past - but issues with `Molecule` and `Structure` would also have far-reaching implications. The unit tests should be expanded upon.

Additionally, more integration tests are required. The OpenMM test is useful, as it tests both structure generation and PSF/PDB format structure. More tests using OpenMM or other simulation packages, perhaps for some more difficult molecules, would be ideal.
