### 2020-03-03 #2132 @jedbrown
@thesketh What is the status of `rdkit`? It fails for me.
```
$ pip install rdkit
ERROR: Could not find a version that satisfies the requirement rdkit (from versions: none)
ERROR: No matching distribution found for rdkit
```
The `pygen_structures` wheel installs, but fails to import due to missing `rdkit`, which seems to be [very old](https://pypi.org/project/rdkit/#history).
