### 2020-03-03 #2132 @thesketh
Unfortunately RDKit has to be installed using Conda because of their installation process. This is an issue with distributing code via different package managers I suppose - sorry about the lack of documentation to that effect, I'll make sure to add it to the README. The conda `environment.yml` in the project root is used by ReadTheDocs to build the documentation, so should be enough to build a minimal environment.
