### 2020-03-03 84bdeee
Changes to README to indicate how to install RDKit, added DOI for VMD paper
