+++ paper.bib
@@ -1,4 +1,4 @@
-
+]
 @article{Brunger1993,
   title = {Computational Challenges for Macromolecular Structure Determination by {{X}}-Ray Crystallography and Solution {{NMRspectroscopy}}},
   author = {Brünger, Axel T. and Nilges, Michael},
@@ -72,6 +72,7 @@
   journaltitle = {Journal of Molecular Graphics},
   volume = {14},
   pages = {33--38}
+  doi = {10.1016/0263-7855(96)00018-5},
 }
 
 @misc{wwPDB2012,
