+++ README.md
@@ -1,41 +1,34 @@
 # pygen-structures
 
-`pygen-structures` (pigeon structures) is a Python utility which allows for the generation of 3 dimensional molecular structures which can be used in molecular dynamics or Monte Carlo simulations.
+`pygen-structures` (pigeon structures) is a Python utility which allows for the generation of 3 dimensional molecular structures which can be used in molecular dynamics or Monte Carlo simulations. Molecules are generated from a list of residues and patches in the format of the CHARMM forcefield, and can be written out as valid PSF and PDB files.
 
-Molecules are generated from a list of residues and patches in the format of the CHARMM forcefield, and can be written out as valid PSF and PDB files.
+The package can be installed using pip (`pip install pygen-structures`), but relies upon _RDKit_, which must be installed using the [conda package manager](https://docs.conda.io/projects/conda/en/latest/) (`conda install -c rdkit rdkit`). _numpy_ is an additional dependency, and both _pytest_ and _OpenMM_ are required to run all the tests.
 
-This should make it significantly easier to perform combinatorial searches on particular sequence lengths and linkages, requiring no manual intervention provided the molecules of interest are reasonably small and the residue/patch definitions already exist in the forcefield.
+In essence, `pygen-structures` aims (eventually) to be a complete [psfgen](https://www.ks.uiuc.edu/Research/vmd/plugins/psfgen/) replacement with more autonomous functionality and without the necessity for an initial 3D structure. This should make it significantly easier to perform combinatorial searches on particular sequence lengths and linkages, requiring no manual intervention provided the molecules of interest are reasonably small and the residue/patch definitions already exist in the forcefield.
 
-In essence, `pygen-structures` aims (eventually) to be a complete [psfgen](https://www.ks.uiuc.edu/Research/vmd/plugins/psfgen/) replacement with more autonomous functions and without the necessity for an initial 3D structure.
+The package can be used as a command line utility, or as a Python library.
 
-`pygen-structures` depends upon _rdkit_ and _numpy_ and can be installed via pip:
+## Command line usage
 
-```
-pip install pygen-structures
-```
-
-Tests can be run with _pytest_, and the only integration test example at present requires `openmm`:
+Command line usage for peptides is incredibly simple, and takes the following form:
 
 ```
-pytest --pyargs pygen_structures
+pygen-structures SEQUENCE -o OUTPUT_PREFIX
 ```
 
-The package can be used as a command line utility, or as a Python library. Command line usage is simple:
+Sequences are specified using the one letter protein code, and terminal patches can be supplied by using hyphens as delimiters (e.g. `NNEU-AFK-CT2`, note that both termini must be supplied). OUTPUT_PREFIX.psf and OUTPUT_PREFIX.pdb are created. If the output prefix is not specified, the PDB file is written to stdout and no PSF file is generated. D-amino acids need only be preceded by a lowercase d.
+
+To generate more complex structures, such as sugars, the residue names should be supplied (hyphen delimited) and the `-u`/`--use-charmm-names` option selected. `--name` and `--segid` control the names given in the COMPND record and the segment id respectively. For raffinose:
 
 ```
 pygen-structures --use-charmm-names AGLC-BFRU-AGAL --patches RAFF 0 1 2 --segid RAFF --name Raffinose -o RAFF
 ```
 
-will produce the structure for the sugar raffinose, saving the PDB file to RAFF.pdb and the PSF file to RAFF.psf. For pure protein structures, the one letter code can be used (the histidine residue type can be set using `--histidine`, but defaults to HSE):
+## Library usage
 
-```
-pygen-structures YEET -o YEET
-```
[…]
