### 2020-03-03 1897681
Removed some hyperbolic language, removed a now-irrelevant sentence and added example for '--patches'
