### 2020-03-03 #2132 @thesketh
I've updated README.md to reflect this information
