### 2020-03-03 eb2b760
Configuration for Travis builds
