+++ pygen_structures/.travis.yml
@@ -1,0 +1,22 @@
+language: python
+python:
+  - "3.6"
+  - "3.7"
+  - "3.8"
+before_install:
+  - sudo apt-get update
+  - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
+  - bash miniconda.sh -b -p $HOME/miniconda
+  - source "$HOME/miniconda/etc/profile.d/conda.sh"
+  - hash -r
+  - conda config --set always_yes yes --set changeps1 no
+  - conda update -q conda
+  - conda info -a
+  - conda create -q -c rdkit -c omnia -n testenv python=$TRAVIS_PYTHON_VERSION rdkit numpy pytest openmm
+  - conda activate testenv
+  - python --version
+install:
+  - python setup.py install
+
+script:
+  - pytest
