### 2020-03-03 #2132 @jedbrown
Okay, I see there is a [thread](https://github.com/rdkit/rdkit/issues/1812) and [pull request](https://github.com/rdkit/rdkit/pull/2690) to make rdkit pip-installable.

I'm getting errors in running the tests, which appear to be due to this:
