### 2020-03-03 0737038
Updated required RDKit version, dropped Python 3.8 (not supported by RDKit)
