+++ setup.py
@@ -28,6 +28,6 @@ setuptools.setup(
         "Operating System :: OS Independent",
         "Typing :: Typed"
     ],
-    python_requires='>=3.6',
+    python_requires='>=3.6,<3.8',
     include_package_data=True,
 )
