+++ .travis.yml
@@ -2,7 +2,6 @@ language: python
 python:
   - "3.6"
   - "3.7"
-  - "3.8"
 before_install:
   - sudo apt-get update
   - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
@@ -12,7 +11,7 @@ before_install:
   - conda config --set always_yes yes --set changeps1 no
   - conda update -q conda
   - conda info -a
-  - conda create -q -c rdkit -c omnia -n testenv python=$TRAVIS_PYTHON_VERSION rdkit numpy pytest openmm
+  - conda create -q -c rdkit -c omnia -n testenv python=$TRAVIS_PYTHON_VERSION 'rdkit>=2018.03' numpy pytest openmm
   - conda activate testenv
   - python --version
 install:
