### 2020-03-03 #2132 @thesketh
For reasons unknown to me, when using Python 3.6 conda was not installing the latest version of the RDKit. I've updated the requirements and .travis.yml, so the builds for 3.6 are now passing.

I was unable to fix a dependency problem with Python 3.8, so I've removed it from the supported versions. I'll reevaluate it when the next RDKit version is released.
