### 2020-03-03 67eb9ad
Added CI and documentation badges
