### 2020-03-04 fef163f
Changes to paper to indicate supported Python versions (3.6, 3.7)
