+++ paper.md
@@ -160,7 +160,7 @@ Further documentation, including a writeup of the types of polymeric structures 
 
 Limited unit and integration tests (using _pytest_) are supplied to test the functionality provided by the package. Tests can be run using `pytest --pyargs pygen_structures`. Tests rely upon _OpenMM_ as an additional dependency.
 
-Any modern Python 3 version should be supported, and pip can be used for installation: `pip install pygen-structures`
+Python 3.6 and 3.7 are supported, and pip can be used for installation: `pip install pygen-structures`.
 
 # Acknowledgements
 
