### 2020-03-10 #2157 @whedon
[REVIEW]: pygen-structures: A Python package to generate 3D molecular structures for simulations using the CHARMM forcefield
**Submitting author:** @thesketh (<a href="http://orcid.org/0000-0002-0502-9596">Travis Paul Hesketh</a>)
**Repository:** <a href="https://github.com/thesketh/pygen-structures" target ="_blank">https://github.com/thesketh/pygen-structures</a>
**Version:** v0.2.4
