+++ paper.bib
@@ -71,8 +71,8 @@
   date = {1996},
   journaltitle = {Journal of Molecular Graphics},
   volume = {14},
-  pages = {33--38}
-  doi = {10.1016/0263-7855(96)00018-5},
+  pages = {33--38},
+  doi = {10.1016/0263-7855(96)00018-5}
 }
 
 @misc{wwPDB2012,
