### 2020-03-10 #2157 @RMeli
I had a first look at the package and have a few comments, none of which I deemed worth of its own issue for now (besides, some issues were already open). I will have a deeper look in the upcoming days, but I thought it might be useful to have some initial comments earlier.

(@richardjgowers if you prefer to see the comments in bulk do let me know and I'll adjust for next time)

## Functionality

### Installation

Installation is easy and described in the `README`. A few comments:
*  Since `RDKit` can only be installed via `conda` I would remove the `requirements.txt` file
* I would add a second `.yml` defining a full environment for testing (so that the user can avoid searching for `openmm` installation instructions)
* Adding a dedicated "Installation" section to the `README` could be beneficial

## Documentation

### Installation

See comments above.

Detailed installation instructions could be added to the main documentation as well (the website is already in place but only the API documentation is present).

### Automated Tests

Automated tests can be easily run with `pytest` and all pass. However, there is no test coverage displayed. Using `pytest-cov` I can see that some modules have rather low coverage (`convenience_functions.py` is at 41%). _I think the test coverage metric should be added and improved for some modules._

(See also  thesketh/pygen-structures#3)

### Functionality Documentation

Documentation for the objects returned by different functions is often missing.

### Example Usage

There is a single usage example. _I think more usage examples should be added_.

It would be useful to describe the different input in more details within the documentation with the help of usage examples.

(See also thesketh/pygen-structures#2; as noted there, the documentation skeleton is already in place and should be relativelystraightforward to add an "Usage Examples" section)

## Misc

I noticed that the CHARMM parameter files are included under `pygen_structures/toppar`. Is there an associated license for such files?
