### 2020-03-11 #2157 @thesketh
Thanks for the preliminary comments, Rocco. I'll address these as soon as I can. I think I can fix the poor coverage of `convenience_functions` by comparing `code_to_mol` and `sequence_to_mol` with a manually created `Molecule`.

The `requirements.txt` file is necessary for PyPI distribution. I could remove RDKit from the requirements and catch the `ImportError` at runtime to produce a more helpful error, but I thought this was a dirtier solution.

The `toppar` distribution is [in the public domain](https://www.charmm.org/ubbthreads/ubbthreads.php?ubb=showflat&Number=33804), but I'll add another `README` in this folder explaining where the files are from, where new versions can be obtained, and what changes have been made from the standard distribution (I've noticed some issues with the D-amino acid parameters, which are fixed in the 0.2.3 branch).

Documentation for return types is missing in some cases because `from __future__ import annotations` is unavailable in Python 3.6, but I'll add `:rtype:` tags where this is the case.
