### 2020-03-11 3c9c8e2
Additional tests, codecov
