+++ .travis.yml
@@ -11,11 +11,12 @@ before_install:
   - conda config --set always_yes yes --set changeps1 no
   - conda update -q conda
   - conda info -a
-  - conda create -q -c rdkit -c omnia -n testenv python=$TRAVIS_PYTHON_VERSION 'rdkit>=2018.03' numpy pytest openmm
+  - conda create -q -c rdkit -c omnia -n testenv python=$TRAVIS_PYTHON_VERSION 'rdkit>=2018.03' numpy pytest pytest-cov codecov openmm
   - conda activate testenv
   - python --version
 install:
   - python setup.py install
-
 script:
-  - pytest
+  - pytest --cov-config .coveragerc --cov=pygen_structures
+after_success:
+  - codecov
