### 2020-03-11 #2157 @RMeli
Thanks for clarifying the situation with `toppar`; the lack of license on their end is a bit  strange... In the light of the forum discussion you linked, what you suggest to do sounds very reasonable.

I'm not well versed with PyPI distribution, but I have a package on PyPI without any `requirements.txt`. I think PyPI deployment only uses `install_requires` in `setup.py`  (see [here](https://packaging.python.org/guides/distributing-packages-using-setuptools/#install-requires)):
> `install_requires` should be used to specify what dependencies a project minimally needs to run. When the project is installed by pip, this is the specification that is used to install its dependencies.

If I `pip`-install `pygen_structures` on a clean environment, the package is installed but neither `numpy` nor `rdkit` are (since `install_requires` is not specified). I think you can remove `requirements.txt`, add `numpy` to `install_requires` and leave `rdkit` installation to the user as described in the documentation (I think the error `ModuleNotFoundError: No module named 'rdkit'` is clear enough). Other reviewers might have different opinions and are probably more experienced with PyPI deployment, so I'd wait for their comments before making any change.
