### 2020-03-11 79b3265
Updates to README to add more examples
