+++ doc/conf.py
@@ -12,6 +12,7 @@
 #
 import os
 import sys
+import sphinx_rtd_theme
 sys.path.insert(0, os.path.abspath('.'))
 sys.path.insert(0, os.path.abspath('../'))
 
@@ -23,7 +24,7 @@ copyright = '2020, Travis Hesketh'
 author = 'Travis Hesketh'
 
 # The full version, including alpha/beta/rc tags
-release = '0.2'
+release = '0.2.2'
 
 
 # -- General configuration ---------------------------------------------------
@@ -33,7 +34,8 @@ release = '0.2'
 # ones.
 extensions = [
 #    "recommonmark",
-    "sphinx.ext.autodoc"
+    "sphinx.ext.autodoc",
+    "sphinx_rtd_theme"
 ]
 
 # Add any paths that contain templates here, relative to this directory.
@@ -50,7 +52,7 @@ exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'alabaster'
+html_theme = 'sphinx_rtd_theme'
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
