+++ README.md
@@ -1,12 +1,21 @@
-# pygen-structures [![Documentation Status](https://readthedocs.org/projects/pygen-structures/badge/?version=latest)](https://pygen-structures.readthedocs.io/en/latest/?badge=latest)[![Build Status](https://travis-ci.org/thesketh/pygen-structures.svg?branch=master)](https://travis-ci.org/thesketh/pygen-structures)
+# pygen-structures
 
-`pygen-structures` (pigeon structures) is a Python utility which allows for the generation of 3 dimensional molecular structures which can be used in molecular dynamics or Monte Carlo simulations. Molecules are generated from a list of residues and patches in the format of the CHARMM forcefield, and can be written out as valid PSF and PDB files.
+[![Documentation Status](https://readthedocs.org/projects/pygen-structures/badge/?version=latest)](https://pygen-structures.readthedocs.io/en/latest/?badge=latest)[![Build Status](https://travis-ci.org/thesketh/pygen-structures.svg?branch=master)](https://travis-ci.org/thesketh/pygen-structures)[![codecov](https://codecov.io/gh/thesketh/pygen-structures/branch/master/graph/badge.svg)](https://codecov.io/gh/thesketh/pygen-structures)
 
-The package can be installed using pip (`pip install pygen-structures`), but relies upon _RDKit_, which must be installed using the [conda package manager](https://docs.conda.io/projects/conda/en/latest/) (`conda install -c rdkit rdkit`). _numpy_ is an additional dependency, and both _pytest_ and _OpenMM_ are required to run all the tests. Python 3.6 and 3.7 are supported.
+`pygen-structures` (pigeon structures) is a Python utility which allows for the generation of 3 dimensional molecular structures which can be used in molecular dynamics or Monte Carlo simulations. Molecules are generated from a list of residues and patches in the format of the CHARMM forcefield, and can be written out as valid PSF and PDB files. The package can be used as a command line utility, or as a Python library.
+
+`pygen-structures` can be installed using pip (`pip install pygen-structures`), but relies upon _RDKit_, which must be installed using the [conda package manager](https://docs.conda.io/projects/conda/en/latest/). For installation instructions, see the 'Installation' section of the readme. Python 3.6 and 3.7 are supported.
 
 In essence, `pygen-structures` aims (eventually) to be a complete [psfgen](https://www.ks.uiuc.edu/Research/vmd/plugins/psfgen/) replacement with more autonomous functionality and without the necessity for an initial 3D structure. This should make it significantly easier to perform combinatorial searches on particular sequence lengths and linkages, requiring no manual intervention provided the molecules of interest are reasonably small and the residue/patch definitions already exist in the forcefield.
 
-The package can be used as a command line utility, or as a Python library.
+## Installation
+
+1. [Install the `conda` package manager](https://docs.anaconda.com/anaconda/install/).
+2. Set up a `conda` environment with the relevant dependencies (or install them in your base distribuion). This can be done with the following command: `conda create -n pygen-structures -c rdkit -c omnia 'python>=3.6' 'rdkit>=2018.3' numpy 'openmm>=7.4' pytest`.
+3. Activate the `conda` environment: `conda activate pygen-structures`
+4. Use `pip` to install `pygen-structures` in this environment: `pip install pygen-structures`.
+5. Installation complete! You will have to activate this environment using `conda activate pygen-structures` each time you want to use it.
+6. Test the installation using `pytest --pyargs pygen_structures`
 
 ## Command line usage
 
@@ -16,18 +25,62 @@ Command line usage for peptides is simple, and takes the following form:
[…]
