### 2020-03-11 deb13bb
Added README for CHARMM toppar distribution
