+++ pygen_structures/toppar/README.md
@@ -1,0 +1,9 @@
+# CHARMM Distribution
+
+This is the July 2019 CHARMM distribution, with all outdated (pre CHARMM35/36) and non-CHARMM format files removed.
+
+Updated files can be downloaded from the [MacKerell lab website](http://mackerell.umaryland.edu/charmm_ff.shtml) when they are made available.
+
+These files are [in the public domain](https://www.charmm.org/ubbthreads/ubbthreads.php?ubb=showflat&Number=33804), but the normal academic conventions apply: please make sure to cite the relevant CHARMM papers if you use the forcefields defined in these files in your work. This helps to ensure further funding.
+
+Any molecules produced by `pygen-structures` will note the topology files which were used in the PSF and PDB headers. The reference for the paper corresponding to each set of topology files can be found on the MacKerell website, as noted above.
