### 2020-03-11 #2157 @thesketh
Thanks for the tip about the `install_requires`, it seems that there's even a page on ["install_requires vs requirements files"](https://packaging.python.org/discussions/install-requires-vs-requirements/#install-requires-vs-requirements-files). I was under the impression that `requirements.txt` was necessary. I think you're probably right and that deleting it is the right thing to do, but I'll hold off for now.

I think I've addressed most of your initial comments now (with the extra information added to README), but I still need to add the information to the main documentation.
