### 2020-03-11 141a3f6
Added numpy to install_requires
