+++ setup.py
@@ -30,4 +30,7 @@ setuptools.setup(
     ],
     python_requires='>=3.6,<3.8',
     include_package_data=True,
+    install_requires=[
+        'numpy',
+    ]
 )
