### 2020-03-11 5392b87
Updates to function return types
