### 2020-03-11 thesketh/pygen-structures#2 @thesketh
More examples added for command line usage in 79b3265580f37aa62d4fd07c393721bb1d6899e6, still need to add examples for library usage. Examples need to be added to RtD documentation.
