### 2020-03-11 thesketh/pygen-structures#3 @thesketh
Additional tests added in 3c9c8e29137c06f5443ecbf8b671d2246a0d19d6, so now over 80% coverage.

Tests to cover command line usage are still necessary.
