### 2020-03-14 thesketh/pygen-structures#6 @amandadumi
Description of Rdkit install without conda in README
Although conda is certainly an easy way, and arguably the easiest way to install rdKit, it is not the **only** way to install rdKit. Instead. Thus line 7, in the [README.md](https://github.com/thesketh/pygen-structures/blob/master/README.md) should be reworded. Instead, it would be more clear to direct the user to the [rdkit install instrcutions](https://www.rdkit.org/docs/Install.html), but suggesting the conda install for your package.

cc:  [openjournals/joss-reviews/2157](https://github.com/openjournals/joss-reviews/issues/2157)
specifically in reference to functionality:installation
