### 2020-03-15 #2157 @amandadumi
I am still making my way through the code, but I had another thought about the handling of the additional testing dependency.  I agree that avoiding having users search for openmm installation instructions would be good.  @RMeli provided this solution, which was nice!

>     * I would add a second `.yml` defining a full environment for testing (so that the user can avoid searching for `openmm` installation instructions)

Another way that this can be done is to define these as extras in the setup.py as described [here](https://setuptools.readthedocs.io/en/latest/setuptools.html#declaring-extras-optional-features-with-their-own-dependencies). I just wanted to mention this since this would avoid an extra file and  could then be installed from pypi as described [here](https://packaging.python.org/tutorials/installing-packages/#installing-setuptools-extras). What do you think, @RMeli? I haven't dealt with this first hand so I'm not sure if your approach has a benefit I am overlooking.
