### 2020-03-15 #2157 @RMeli
@amandadumi I think that would be a good idea, but it seems to me that in order to [install OpenMM](http://docs.openmm.org/latest/userguide/application.html#installing-openmm) `conda` is needed (with the `-c omnia -c conda-forge` channels).
