### 2020-03-15 thesketh/pygen-structures#3 @RMeli
Any short-term plans to bump code coverage even higher?
