### 2020-03-15 #2157 @richardjgowers
Iirc there was an effort to get a minimal openmm working on conda-forge...
the hang up was cuda dependencies
