### 2020-03-15 #2157 @richardjgowers
By @jaimergp

On Sun, Mar 15, 2020 at 20:06, Richard Gowers <richardjgowers@gmail.com>
wrote:

> Iirc there was an effort to get a minimal openmm working on conda-forge...
> the hang up was cuda dependencies
>
> On Sun, Mar 15, 2020 at 20:03, Rocco Meli <notifications@github.com>
> wrote:
>
>> @amandadumi <https://github.com/amandadumi> I think that would be a good
>> idea, but it seems to me that in order to install OpenMM
>> <http://docs.openmm.org/latest/userguide/application.html#installing-openmm>
>> conda is needed (with the -c omnia -c conda-forge channels).
>>
>> —
>> You are receiving this because you were mentioned.
>>
>>
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/2157#issuecomment-599257354>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/ACGSGB4X4PQU3WJD7KE5PLDRHUYCDANCNFSM4LFDODLA>
>> .
>>
>
