### 2020-03-15 #2157 @RMeli
@thesketh I didn't have enough time to dig deep into the code, but I have some comments about the paper below.

## Software Paper

### Statement of Need

A big chunk of Fig. 1 details "planned future work". This gives the impression that `pygen-structures` is incomplete and there is still a lot of work to do. I'm wondering if it  would be better to focus on what the program does instead of outlining a lot of planned things that could be far away in the future. Generating coordinates for missing residues is a big project in itself and you already underline this limitation in the conclusions. @richardjgowers might have some advice on this.

### Summary

I'm not sure about the term "arbitrary molecules", since you can only generate molecules from CHARMM residues and patches.

### Problems with Current Workflow

* [ ] Is there a way to fit the content of the PDB file withing the paper margins?

I think the problems with current workflows are very well illustrated. Despite only two packages are discussed (RDKit and MarvinSketch), I think similar problems will arise with other packages.

### Future Work

I think it's great that you outline the limitation of  `pygen-structures` compared to `psfgen` and possible big  additions to the functionality of the software. However, I think that the documentation of the current features should be complete and not left as  future work.

### Dependencies

* [ ] Citations for Numpy is missing.
* [ ] Unit and  integration tests should be complete and not _limited_.

### Acknowledgements

* [ ]  Is there any funding body to be acknowledge?

### Misc

* [ ] It's worth adding more details to your affiliation (department, address)?
* [ ]  `psfgen` is inconsistently written in italic
