### 2020-03-16 #2157 @jaimergp
Hi! Yes, we have a functional `openmm` recipe on `conda-forge`, but still not publicly available (it's on the `testing` label). We have to sort out how to deal with potentially unsupported OpenCL implementations (check https://github.com/conda-forge/cfep/issues/17). In the meantime, the `omnia` channel is needed, but as you can see in the issue this is subject to change in the near-term future.
