### 2020-03-16 9ef3298
Added tests for command line usage
