+++ pygen_structures/test/test_command_line.py
@@ -1,0 +1,210 @@
+import os
+import warnings
+from pygen_structures.convenience_functions import (
+    load_charmm_dir,
+    pdb_to_mol
+)
+from pygen_structures import __main__ as cmd_interface
+
+def test_arg_parsing():
+    argv = ["HEY", "-o", "HEY_out", "--histidine", "HSP"]
+    args = cmd_interface.parse_args(argv)
+
+    assert(args.sequence == "HEY")
+    assert(args.segid == "PROT")
+    assert(args.patches == None)
+    assert(args.toppar == None)
+    assert(args.verify == True)
+    assert(args.output == "HEY_out")
+    assert(args.histidine == "HSP")
+    assert(args.use_charmm_names == False)
+
+    argv = [
+        "-u", "HSE-TRP-LYS", "-o", "HWK", "--patches", "CT2", "LAST",
+        "-v", "--segid", "HWK"
+    ]
+    args = cmd_interface.parse_args(argv)
+    assert(args.sequence == "HSE-TRP-LYS")
+    assert(args.segid == "HWK")
+    assert(args.patches == ["CT2", "LAST"])
+    assert(args.toppar == None)
+    assert(args.verify == False)
+    assert(args.output == "HWK")
+    assert(args.histidine == "HSE")
+    assert(args.use_charmm_names == True)
+
+def test_molecule_creation():
+    argv = [
+        "-u", "AGLC-BFRU-AGAL", "-o", "RAFF",
+        "--patches", "RAFF", "0", "1", "2",
+        "--segid", "RAFF", "--name", "Raffinose"
+    ]
+    cmd_interface.main(argv)
+
+    assert(os.path.exists("RAFF.psf"))
+    assert(os.path.exists("RAFF.pdb"))
+
+    rtf, prm = load_charmm_dir()
+    with warnings.catch_warnings():
+        warnings.simplefilter('ignore')
+        molecule = pdb_to_mol("RAFF.pdb", rtf, patches={"RAFF": (0, 1, 2)})
+
+    assert(molecule.name == "Raffinose")
+    assert(molecule.segment == "RAFF")
+    assert(molecule.check_parameters(prm))
+
+    ref_atoms = {
+        (0, 'C1'),
+        (0, 'H1'),
+        (0, 'O1'),
+        (0, 'C5'),
+        (0, 'H5'),
+        (0, 'O5'),
+        (0, 'C2'),
+        (0, 'H2'),
+        (0, 'O2'),
+        (0, 'HO2'),
+        (0, 'C3'),
+        (0, 'H3'),
+        (0, 'O3'),
+        (0, 'HO3'),
+        (0, 'C4'),
+        (0, 'H4'),
+        (0, 'O4'),
+        (0, 'HO4'),
+        (0, 'C6'),
+        (0, 'H61'),
+        (0, 'H62'),
+        (1, 'O5'),
+        (1, 'C2'),
+        (1, 'C5'),
+        (1, 'H5'),
+        (1, 'C6'),
+        (1, 'H61'),
+        (1, 'H62'),
+        (1, 'O6'),
+        (1, 'HO6'),
+        (1, 'C1'),
+        (1, 'H11'),
+        (1, 'H12'),
+        (1, 'O1'),
+        (1, 'HO1'),
+        (1, 'C3'),
+        (1, 'H3'),
[…]
