### 2020-03-16 1854775
Improved test coverage for command line usage
