### 2020-03-16 6e9146e
Fixes to regex patterns
