### 2020-03-16 9823cc4
Changes to paper and readme
