+++ paper.bib
@@ -1,4 +1,4 @@
-]
+
 @article{Brunger1993,
   title = {Computational Challenges for Macromolecular Structure Determination by {{X}}-Ray Crystallography and Solution {{NMRspectroscopy}}},
   author = {Brünger, Axel T. and Nilges, Michael},
@@ -11,7 +11,7 @@
   url = {https://www.cambridge.org/core/journals/quarterly-reviews-of-biophysics/article/computational-challenges-for-macromolecular-structure-determination-by-xray-crystallography-and-solution-nmrspectroscopy/1EF4F138E8B6A1BC6924766F36BC52F6},
   urldate = {2020-02-25},
   abstract = {Macromolecular structure determination by X-ray crystallography and solution NMR spectroscopy has experienced unprecedented growth during the past decade.},
-  file = {/users/gvb14176/Zotero/storage/S99DQX2X/Brünger and Nilges - 1993 - Computational challenges for macromolecular struct.pdf;/users/gvb14176/Zotero/storage/GCJG2ZE5/1EF4F138E8B6A1BC6924766F36BC52F6.html},
+  file = {/home/travis/Zotero/storage/S99DQX2X/Brünger and Nilges - 1993 - Computational challenges for macromolecular struct.pdf;/home/travis/Zotero/storage/GCJG2ZE5/1EF4F138E8B6A1BC6924766F36BC52F6.html},
   langid = {english},
   number = {1}
 }
@@ -33,6 +33,45 @@
   number = {25}
 }
 
+@article{numpy,
+  title = {The {{NumPy Array}}: {{A Structure}} for {{Efficient Numerical Computation}}},
+  shorttitle = {The {{NumPy Array}}},
+  author = {van der Walt, Stefan and Colbert, S. Chris and Varoquaux, Gael},
+  date = {2011-03},
+  journaltitle = {Computing in Science Engineering},
+  volume = {13},
+  pages = {22--30},
+  issn = {1558-366X},
+  doi = {10.1109/MCSE.2011.37},
+  abstract = {In the Python world, NumPy arrays are the standard representation for numerical data and enable efficient implementation of numerical computations in a high-level language. As this effort shows, NumPy performance can be improved through three techniques: vectorizing calculations, avoiding copying data in memory, and minimizing operation counts.},
+  file = {/home/travis/Zotero/storage/S92DK667/van der Walt et al. - 2011 - The NumPy Array A Structure for Efficient Numeric.pdf},
+  keywords = {Arrays,Computational efficiency,data structures,Finite element methods,high level language,high level languages,mathematics computing,numerical analysis,Numerical analysis,numerical computation,numerical computations,numerical data,NumPy,numpy array,Performance evaluation,programming libraries,Python,Python programming language,Resource management,scientific programming,Vector quantization},
+  note = {Conference Name: Computing in Science Engineering},
+  number = {2},
+  options = {useprefix=true}
+}
+
+@article{openmm,
+  title = {{{OpenMM}} 7: {{Rapid}} Development of High Performance Algorithms for Molecular Dynamics},
+  shorttitle = {{{OpenMM}} 7},
+  author = {Eastman, Peter and Swails, Jason and Chodera, John D. and McGibbon, Robert T. and Zhao, Yutong and Beauchamp, Kyle A. and Wang, Lee-Ping and Simmonett, Andrew C. and Harrigan, Matthew P. and Stern, Chaya D. and Wiewiora, Rafal P. and Brooks, Bernard R. and Pande, Vijay S.},
+  date = {2017-07-26},
+  journaltitle = {PLOS Computational Biology},
+  shortjournal = {PLOS Computational Biology},
+  volume = {13},
[…]
