### 2020-03-16 thesketh/pygen-structures#6 @thesketh
True, I've just never thought to install RDKit any other way! I've fixed this in 9823cc47dac159d12054eebd398ba3c9782804db
