### 2020-03-16 thesketh/pygen-structures#3 @thesketh
I've added tests for the command-line portion of the software, so coverage is ~91% now
