+++ requirements.txt
@@ -1,2 +1,0 @@
-rdkit>=2018.03
-numpy
