+++ pygen_structures/version.py
@@ -2,5 +2,5 @@
 Just a version string! Changing the version here will update it
 everywhere else it is used in the project.
 """
-version = "0.2.2"
+version = "0.2.3"
 year = "2020"
