### 2020-03-16 c4d0db1
Update MANIFEST.in
