### 2020-03-16 #2157 @amandadumi
@thesketh and @Rmeli, this is clear now. I do not use openmm and was didn't realize the [openmm pypi project](https://pypi.org/project/openmm/) was simply for a user-friendly interface.

I was also able to go through the paper and had some general comments, many of mine echo what had previously been said, but I had a few specifics that applied to the most recent article proof. I don't think the call to whedon to include the  updated flowchart worked though.

**Software paper**

A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
- [ ] wording of figure 1 and title of Problems with current work
    - Above in the diagram, what `pygen-structures` can do is referred to as current work, which is fine. However, when current workflows is used in the section title 'Problems with current workflows'. Additionally, this sections main takeaway isn't the problems of others, but more so is highlighting where `pygen-structures` out-performs previous workflow. The section is really demonstrative and beneficial, but the section naming is a confusing. Some suggestions of section title that reflects the contents and is avoid the confusing use of 'current' is 'Demonstrating the benefit of `pygen-structures` ' or simply 'Comparison of workflows'.

State of the field: Do the authors describe how this software compares to other commonly-used packages?
- [ ] you mention other chemical drawing packages, without citing which would behave like this one. I assume then it is all popular ones such as Avogadro, Pymol? (* I see you discussed this above*)  I don't think this needs explicitly mentioned in text (*or even have examples provided as I agree the examples would add unnecessary length to the document.*), but just cite the codes you know behave similarly.

Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
- [ ] Overall the structure is great and the language  is clear and concise.I have a few small recommendations:

Small recommendations:
- In the second paragraph in the **Statement of Need** section, the sentence "Tools which are capable of..." makes it seem like this is an unavoidable problem. Instead being specific to "Existing tools..." would avoid this.
- under the Problems with current workflows, a sentence starts "In the RDKit,", which is awkward.  An alternative would simple removing the "the", but I think it would be better to say "RDKit has the capability to generate simple protein sequences..."
-  The structure of the first sentence of the summary feels bulky. Additionally, some slight rewording can avoid the use of an ambiguous "them". Suggested rewording:
 " `pygen-structures` is an open source (3-clause BSD license) Python library to generating 3 dimensional structures for molecules from the CHARMM forcefield based on empirical data (Riniker & Landrum (2015), as implemented in the RDKit, (“RDKit,”2020)).

References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)?
- [ ] the [SciPy ecosystem page](https://www.scipy.org/citing.html) suggests two citations for NumPy: the van der Walt *et al.* paper you've included but also Travis E. Oliphant. A guide to NumPy, USA: Trelgol Publishing, (2006).
