### 2020-03-17 #2157 @dotsdl
Hi @thesketh, here are my review notes. Great work on this submission; I believe this to be a good fit for JOSS. I am satisfied that much of the criteria for submission are met, and so the notes below focus heavily on items that I think should be addressed yet.

## High-level notes

1. Nice tool! I find the CHARMM forcefield tooling to be particularly difficult to work with, so this is a welcome addition to the Python tooling for this area.
2. A lot of the functionality appears to be planned, but is not currently implemented.
3. The project is less than a month old, with a single contributor. This is not necessarily a problem, but it is an indication to me that this library may benefit from more development before publishing in JOSS.
4. This tool looks very useful to CHARMM users who need to work with small amino acid sequences. It would be of limited utility outside of either of those needs. This is not necessarily a problem, but a note on the scope of the community it serves.

## Review (2020.03.16)

### Functionality

1. Only appears to work for short segments (<15 residue from my ad-hoc tests). This should probably be made more explicit.
   - appears to be addressed in README.
2. The tool works quite well for short segments, and is much easier to use than other tools that can give similar results.

### Documentation

1. Essentially all documentation is in the README. What is present is a small but good set of examples on usage of the command-line tool.
2. There isn't really any documentation on how to use the library components, or what the API is.
    - I would like there to be some API components called out for those that would like to use the library as part of larger frameworks, not just as a command-line tool.
    - the paper gives one example of library component usage. It would be better to see some of this in the library docs themselves as well.

### Software paper

1. Only appears to work for short (<15 residue) segments. This should probably be made more explicit in the paper.
   - appears to be addressed in README.
