### 2020-03-17 f6c3ac3
Changes to address reviewer comments
