+++ paper.bib
@@ -51,6 +51,20 @@
   options = {useprefix=true}
 }
 
+@book{numpyguide,
+  title = {Guide to {{NumPy}}},
+  author = {Oliphant, Travis},
+  date = {2015},
+  edition = {Second},
+  publisher = {{CreateSpace Independent Publishing Platform}},
+  url = {http://doi.org/10.5555/2886196},
+  urldate = {2020-03-17},
+  abstract = {This is the second edition of Travis Oliphant's A Guide to NumPy originally published electronically in 2006. It is designed to be a reference that can be used by practitioners who are familiar with Python but want to learn more about NumPy and related tools. In this updated edition, new perspectives are shared as well as descriptions of new distributed processing tools in the ecosystem, and how Numba can be used to compile code using NumPy arrays. Travis Oliphant is the co-founder and CEO of Continuum Analytics. Continuum Analytics develops Anaconda, the leading modern open source analytics platform powered by Python. Travis, who is a passionate advocate of open source technology, has a Ph.D. from Mayo Clinic and B.S. and M.S. degrees in Mathematics and Electrical Engineering from Brigham Young University. Since 1997, he has worked extensively with Python for computational and data science. He was the primary creator of the NumPy package and founding contributor to the SciPy package. He was also a co-founder and past board member of NumFOCUS, a non-profit for reproducible and accessible science that supports the PyData stack. He also served on the board of the Python Software Foundation.},
+  file = {/home/travis/Zotero/storage/6NQVV58M/2886196.html},
+  isbn = {978-1-5173-0007-4},
+  pagetotal = {364}
+}
+
 @article{openmm,
   title = {{{OpenMM}} 7: {{Rapid}} Development of High Performance Algorithms for Molecular Dynamics},
   shorttitle = {{{OpenMM}} 7},
