+++ paper.md
@@ -20,19 +20,19 @@ bibliography: paper.bib
 
 Forcefields used in molecular dynamics (MD) or Monte Carlo simulations (such as the CHARMM forcefield, @charmm) require a great deal of information in order to ensure that the correct harmonic force constants are chosen for particular bonds or angles. As such, protein structure files (PSF files), containing information about connectivity and relevant parameters, are often used by simulation packages to contain this information. These files are typically generated using _psfgen_, [@psfgen] a Tcl package bundled with _VMD_ [@vmd] which requires a fully annotated Protein Data Bank (PDB) format file [@wwPDB2012] with correct names for atoms and residues, as defined in the forcefield. This annotation is challenging to do in an automated fashion. For structures from the Protein Data Bank itself, this is expected to some degree, as experimentally resolved structures are often missing the coordinates of certain residues [@Brunger1993] and can contain bound ligands which may not be parameterised in the forcefield. At present, the normal solution to this problem is to download a structure from the database, manually fix residue and atom names, and generate a PSF file using psfgen. In other words, obtaining a 3D structure, annotating the structure and generating an input file in the correct format for a simulation package are treated as separate tasks.
 
-For simulations involving combinatorial searches of smaller sequences, however, the level of human intervention required can be far greater. There may be no experimentally determined structures for a given sequence, so these must be generated. Tools which are capable of generating 3D conformations of such molecules must be applicable more generally than molecular dynamics forcefields, and as a result are far less specific about the identities of particular atoms. As such, these packages often leave residue and atom names labelled as unknowns or element symbols. Additionally, if structures are generated based only on connectivity or drawn by the user, it can be challenging to ensure that tetrahedral chirality (the _handedness_ of carbon centres with four different substituents) is properly respected. A structure generation method for small molecules which is aware of the forcefield is necessary in order to automate this process. The workflow which `pygen-structures` hopes to automate is shown in Figure 1.
+For simulations involving combinatorial searches of smaller sequences, however, the level of human intervention required can be far greater. There may be no experimentally determined structures for a given sequence, so these must be generated. Existing structure generation tools must be applicable to more general chemistry than molecular dynamics forcefields, and as a result are far less specific about the identities of particular atoms. As such, these packages often leave residue and atom names labelled as unknowns or element symbols. Additionally, if structures are generated based only on connectivity or drawn by the user, it can be challenging to ensure that tetrahedral chirality (the _handedness_ of carbon centres with four different substituents) is properly respected. A structure generation method for small molecules which is aware of the forcefield is necessary in order to automate this process. The workflow which `pygen-structures` hopes to automate is shown in Figure 1. As with all structure generation methods, applicability is limited to cases where secondary structure is unimportant.
 
 ![An example workflow used to prepare structures for molecular dynamics or Monte Carlo simulation. For smaller polymeric structures, `pygen-structures` can be used to prevent the need for significant manual intervention.](Figures/WorkflowFlowchart.png)
 
 # Summary
 
-`pygen-structures` is an open source (3-clause BSD license) Python library to work with molecules from the CHARMM forcefield, generating 3 dimensional structures for them based on empirical data (@Riniker2015, as implemented in the _RDKit_, @rdkit), writing out standard PSF and PDB files. The package contains convenience functions for generating molecules from a collection of CHARMM residues and patches, or from one letter amino acid codes (usable by calling `pygen-structures` as a command line application) and a series of classes and functions for representing and manipulating CHARMM data. The chirality of tetrahedral centres is set using internal coordinate data from the residue topology file.
[…]
