### 2020-03-17 #2157 @thesketh
Thank you both for your comments!

@amandadumi I think I've addressed most of these now. I haven't changed the sentence containing "the RDKit" as this is how the package is referred to in [its documentation.](https://www.rdkit.org/docs/GettingStartedInPython.html). I read it as "the Rational Discovery Kit" in my head.

Yes, in the past I've encountered similar behaviour from different 2D and 3D chemical drawing packages. I don't have access to a Windows/Mac computer to check how ChemDraw behaves, and I haven't tried to use PyMol before, but I'll try that and add the citations for these packages and Avogadro.

@dotsdl I submitted the paper at this stage because I believe the introduced functionality - in particular the ability to generate annotated structures for polysaccharides, lipids or protein-sugar linkages - is unique. Other additions would increase usability, but are not limiting progress towards an automated workflow. Much of the planned functionality is "would be nice to have", but is covered by existing software packages (e.g. Modeller to assign positions for missing residues, psfgen to handle protein structures from the protein databank and fill missing atoms, VMD's solvate plugin to handle creation of simulation boxes, and the martinize.py script to coarse grain protein structures for the MARTINI forcefield).

The project has been on GitHub for around a month, but I've been working on it piecemeal for a few months. I am still the only contributor so far,  but I'd like to encourage further contribution. You are right that the package is tied in to the CHARMM ecosystem, do you think I should make this more clear?

I'm working on addressing the lack of documentation for library usage, and I've updated the paper to be more explicit about the requirement for structures to be small.
