### 2020-03-18 #2157 @amandadumi
@thesketh ha! I have never noticed that in RDKit documentation! Thanks for explaining, that makes sense then.
