### 2020-03-19 #2157 @amandadumi
## Documentation
@thesketh, your README is very well done and clear.  I am also very happy to see the use of typing in your code! I found it very useful.  Additionally, I found that your function commenting was thorough!

Two small issues:
- [ ] Echoing the point made above, I would like clearer examples of using this package as a Python library, especially since the capability is stated in the README. Which relates to issue thesketh/pygen-structures/issues/2. I realize you're working on this, but I wanted to explain why I am waiting to check that box is all.

- [ ] I like the inclusion of module descriptions as a bird eye view of the modules purpose, but I find it strange these appear at the bottom of the documentation page. Is this according to a specific standard? Otherwise, the descriptions would be more helpful at the top of the specific module. Specifically:
  - https://pygen-structures.readthedocs.io/en/latest/source/pygen_structures.mol_containers.html#
  - https://pygen-structures.readthedocs.io/en/latest/source/pygen_structures.mol_containers.html#module-pygen_structures.mol_containers.structure

# Code
- [ ] The smallest fix: in `charmm_containers.py`, line 44 don't has a " instead of  '.
