### 2020-03-19 #2157 @RMeli
@thesketh Thanks for addressing our comments on the paper (I think the limitation on the number of residues, outlined by @dotsdl, is still missing?). The only "issue" I have left with the paper is the overflow of the code with respect to the text margins. @richardjgowers will there be any editing after the review, @thesketh has to take care of it, or it is OK as it is?
