### 2020-03-19 #2132 @arfon
> @arfon I noticed my name appears three times in the list of reviews. Would it be possible to remove the first two entries (entry 1127 is the most up-to-date)? Thanks.

Not sure I follow sorry?
