### 2020-03-19 #2132 @RMeli
Sorry, I wasn't clear. My name (GitHub ID) appears three times in the [this list of people [that] have already agreed to review for JOSS](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822). I was wondering if the first two spurious entries could be removed.
