### 2020-03-19 #2132 @arfon
> Sorry, I wasn't clear. My name (GitHub ID) appears three times in the this list of people [that] have already agreed to review for JOSS. I was wondering if the first two spurious entries could be removed to avoid confusion.

Done.
