+++ paper.bib
@@ -1,4 +1,23 @@
 
+@article{avogadro,
+  title = {Avogadro: An Advanced Semantic Chemical Editor, Visualization, and Analysis Platform},
+  shorttitle = {Avogadro},
+  author = {Hanwell, Marcus D. and Curtis, Donald E. and Lonie, David C. and Vandermeersch, Tim and Zurek, Eva and Hutchison, Geoffrey R.},
+  date = {2012-08-13},
+  journaltitle = {Journal of Cheminformatics},
+  shortjournal = {J Cheminform},
+  volume = {4},
+  pages = {17},
+  issn = {1758-2946},
+  doi = {10.1186/1758-2946-4-17},
+  url = {https://doi.org/10.1186/1758-2946-4-17},
+  urldate = {2020-03-23},
+  abstract = {The Avogadro project has developed an advanced molecule editor and visualizer designed for cross-platform use in computational chemistry, molecular modeling, bioinformatics, materials science, and related areas. It offers flexible, high quality rendering, and a powerful plugin architecture. Typical uses include building molecular structures, formatting input files, and analyzing output of a wide variety of computational chemistry packages. By using the CML file format as its native document type, Avogadro seeks to enhance the semantic accessibility of chemical data types.},
+  file = {/home/travis/Zotero/storage/RH7CCSG5/Hanwell et al. - 2012 - Avogadro an advanced semantic chemical editor, vi.pdf},
+  langid = {english},
+  number = {1}
+}
+
 @article{Brunger1993,
   title = {Computational Challenges for Macromolecular Structure Determination by {{X}}-Ray Crystallography and Solution {{NMRspectroscopy}}},
   author = {Brünger, Axel T. and Nilges, Michael},
@@ -31,6 +50,13 @@
   eprinttype = {pmid},
   keywords = {calmodulin,cold-shock protein A,empirical force field,fatty acid binding protein,G protein B1,J-coupling,lysozyme,molecular mechanics,residual dipolar coupling,ubiquitin},
   number = {25}
+}
+
+@software{chemoffice,
+  title = {The {{ChemOffice Suite}}},
+  author = {{PerkinElmer Informatics}},
+  date = {2019},
+  version = {19}
 }
 
 @article{numpy,
@@ -95,6 +121,13 @@
   urldate = {2020-02-25}
 }
 
+@software{pymol,
+  title = {The {{PyMOL Molecular Graphics System}}},
+  author = {{Schrödinger LLC}},
+  date = {2019},
+  version = {2.3}
+}
+
 @article{pytest,
   title = {Pytest 5.0},
   author = {Krekel, Holger and Oliveira, Bruno and Pfannschmidt, Ronny and Bruynooghe, Floris and Laugher, Brianna and Bruhin, Florian},
@@ -102,9 +135,10 @@
   url = {https://github.com/pytest-dev/pytest}
 }
 
-@misc{rdkit,
+@online{rdkit,
   title = {{{RDKit}}: {{Open}}-{{Source Cheminformatics Software}}},
   shorttitle = {{{RDKit}}},
+  author = {{The RDKit Contributors}},
   url = {https://www.rdkit.org/},
   urldate = {2020-02-25}
 }
