### 2020-03-23 1a0a145
More paper formatting fixes
