### 2020-03-23 #2157 @thesketh
Thank you both for your comments.

@amandadumi Well spotted with " instead of '! That will teach me not to use find and replace to fix inconsistent string quotes. As for the module descriptions appearing at the bottom, I'm unsure why they're in the wrong place but I'll try to adjust it. These are autogenerated from the code comments, so should be fixed when I add the documentation for use as a library function. I've added citations for the other drawing packages as you requested.

@RMeli I've added an explicit reference to the maximum number of residues, and I fixed the long lines in the code examples by putting in line continuation characters (and elipses on the following line). I think this should be clear enough, but I can update this if you think there's a better way to show that the lines are wrapped or if you think it should be mentioned somewhere in the text.

I think I've now made all the recommended changes to the paper itself, I'll try to get the library documentation posted early this week.
