### 2020-03-23 #2157 @amandadumi
@thesketh  All you will need to do is to go into each of the pages content description within the [doc/source/](https://github.com/thesketh/pygen-structures/tree/master/doc/source) folder. The order of the content for each page is dictated by these files. Moving the  "Module Contents" section to the top should do the trick!
