+++ doc/index.rst
@@ -3,12 +3,13 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-pygen-structures documentation
+pygen-structures Documentation
 ==============================
 
 .. toctree::
    :maxdepth: 1
    :caption: Contents:
+
    installation
    command_line_usage
    library_usage
