### 2020-03-25 c0ae1c2
Some work on the docs, updates to docstrings
