+++ doc/source/pygen_structures.rst
@@ -1,8 +1,5 @@
-pygen\_structures package
-=========================
-
-Module contents
----------------
+API Reference
+=============
 
 .. automodule:: pygen_structures
    :members:
@@ -19,34 +16,7 @@ Subpackages
 Submodules
 ----------
 
-pygen\_structures.charmm\_containers module
--------------------------------------------
+.. toctree::
 
-.. automodule:: pygen_structures.charmm_containers
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pygen\_structures.convenience\_functions module
------------------------------------------------
-
-.. automodule:: pygen_structures.convenience_functions
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pygen\_structures.tests module
-------------------------------
-
-.. automodule:: pygen_structures.tests
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pygen\_structures.version module
---------------------------------
-
-.. automodule:: pygen_structures.version
-   :members:
-   :undoc-members:
-   :show-inheritance:
+   pygen_structures.charmm_containers
+   pygen_structures.convenience_functions
