+++ doc/source/pygen_structures.mol_containers.molecule.rst
@@ -1,0 +1,7 @@
+pygen\_structures.mol\_containers.molecule module
+=================================================
+
+.. automodule:: pygen_structures.mol_containers.molecule
+   :members:
+   :undoc-members:
+   :show-inheritance:
