+++ doc/source/pygen_structures.mol_containers.rst
@@ -1,9 +1,6 @@
 pygen\_structures.mol\_containers package
 =========================================
 
-Module contents
----------------
-
 .. automodule:: pygen_structures.mol_containers
    :members:
    :undoc-members:
@@ -12,26 +9,8 @@ Module contents
 Submodules
 ----------
 
-pygen\_structures.mol\_containers.atom module
----------------------------------------------
+.. toctree::
 
-.. automodule:: pygen_structures.mol_containers.atom
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pygen\_structures.mol\_containers.molecule module
--------------------------------------------------
-
-.. automodule:: pygen_structures.mol_containers.molecule
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pygen\_structures.mol\_containers.structure module
---------------------------------------------------
-
-.. automodule:: pygen_structures.mol_containers.structure
-   :members:
-   :undoc-members:
-   :show-inheritance:
+   pygen_structures.mol_containers.atom
+   pygen_structures.mol_containers.molecule
+   pygen_structures.mol_containers.structure
