+++ doc/source/pygen_structures.mol_containers.atom.rst
@@ -1,0 +1,7 @@
+pygen\_structures.mol\_containers.atom module
+=============================================
+
+.. automodule:: pygen_structures.mol_containers.atom
+   :members:
+   :undoc-members:
+   :show-inheritance:
