+++ doc/source/pygen_structures.convenience_functions.rst
@@ -1,0 +1,7 @@
+pygen\_structures.convenience\_functions module
+===============================================
+
+.. automodule:: pygen_structures.convenience_functions
+   :members:
+   :undoc-members:
+   :show-inheritance:
