+++ doc/library_usage.rst
@@ -1,11 +1,13 @@
 Library Usage
 ==================
 
+
 Representations
 ---------------
 
 pygen-structures contains classes to represent CHARMM data, molecular (and
 molecular-like) structures and atoms.
+
 
 Representing CHARMM File Types
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -43,7 +45,9 @@ simulations, the representation of the ``.rtf`` files is complete,
 but the representation of ``.prm`` files is limited to verifying that
 the parameters are present rather than probing their values.
 
-**Residue Files**
+
+Residue Files
+"""""""""""""
 
 CHARMM ``.rtf`` files are represented using the
 :class:`~pygen_structures.charmm_containers.CHARMMResidueTopologyFile` class,
@@ -69,7 +73,9 @@ Daltons are stored in ``dict`` attribute
 :attr:`~pygen_structures.charmm_containers.CHARMMResidueTopologyFile.masses`,
 where keys are the atom types.
 
-**Parameter Files**
+
+Parameter Files
+"""""""""""""""
 
 CHARMM ``.prm`` files are represented using the
 :class:`~pygen_structures.charmm_containers.CHARMMParameterFile` class.
@@ -103,7 +109,8 @@ specified in reverse order without altering the relevant parameters.
 Representing Residues and Patches
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-**The Difference Between Definitions and Residues**
+The Difference Between Definitions and Residues
+"""""""""""""""""""""""""""""""""""""""""""""""
 
 CHARMM residues as presented in ``.rtf`` files are an abstract residue
 definition, rather than physical residues. As such, these lack an associated
@@ -114,26 +121,210 @@ class.
 
 CHARMM residues in an actual structure are not abstract, and refer to real
 residues in a physical molecule. As such, these _do_ have an associated index
-and their degrees of freedom refer to atom IDs - tuples of
+and their degrees of freedom refer to *atom IDs* - tuples of
 ``(residue_index: int, atom_name: str)``. This is to allow for easier removal of
 dangling bonds and easier application of patches. These residues are represented
 using the :class:`~pygen_structures.charmm_containers.CHARMMResidue` class.
 Patch definitions can only be applied to real residues.
 
-**CHARMM Residue Definitions**
+The patch definitions, instead of the atom IDs used in the residues, use
+*atom references*. These take the same format as atom IDs, but the residue index
+is replaced by an integer referring to the order the residues are provided to
+the patch. For the ``RAFF`` patch, which creates the sugar raffinose from
+``[AGLC, BFRU, AGAL]``, atoms in ``AGLC`` would take the form
+``(0, atom_id: str)`` (``1`` for ``BFRU``, ``2`` for ``AGAL``). Patch
+definitions are represented using the
+:class:`~pygen_structures.charmm_containers.CHARMMPatchResidueDefinition`
+class.
 
-**CHARMM Residues**
+CHARMM Residue Definitions
+""""""""""""""""""""""""""
+
+CHARMM residue definitions
+(:class:`~pygen_structures.charmm_containers.CHARMMResidueDefinition`)
+represent all of the data from a ``RESI`` block in a ``.rtf`` file.
+This includes:
+
+-  the ``name``
+-  a ``list`` of atoms from the ATOM records (represented as
+   ``(atom_name: str, atom_type: str, partial_charge: float)``)
+-  a ``list`` of bonds - ``tuple`` of 2 atom_names - from BOND/DOUB records.
+   Bonds in DOUB records appear twice.
+-  a ``list`` of impropers – ``tuple`` of 4 atom_name from IMPR records.
+-  a ``list`` of cross_maps – ``tuple`` of 8 atom_name from CMAP records
[…]
