### 2020-03-27 #2157 @thesketh
Thanks for the tip, @amandadumi! I ended up re-generating the RST files using `sphinx-apidoc`, the `-M`flag put them in the correct order.

[I believe the docs on ReadTheDocs have a fairly comprehensive introduction to library usage on them now.](https://pygen-structures.readthedocs.io/en/latest/index.html)
