### 2020-03-27 a13ccef
Updated version in doc/conf.py
