+++ doc/conf.py
@@ -24,7 +24,7 @@ copyright = '2020, Travis Hesketh'
 author = 'Travis Hesketh'
 
 # The full version, including alpha/beta/rc tags
-release = '0.2.2'
+release = '0.2.3'
 
 
 # -- General configuration ---------------------------------------------------
