### 2020-03-27 accd173
Slight update to formatting in docs [skip ci]
