### 2020-03-28 #2157 @RMeli
@thesketh the documentation is looking great now, well done!

A few comments below:

---

> --name and --segid control the names given in the COMPND record and the segment ID (4 character max) respectively.

When I try to use a `--segid` longer than 4 characters, the code executes normally without warning. I think this is ok (since user should know this limitation), but maybe it is worth adding a warning that the name will be truncated just in case?

---

Under the **Library Usage - Basic Usage** section there is an error in the 4th code block:
```
patches = {'RAFF', [0, 1, 2]}  # Should be a dictionary: {'RAFF': [0, 1, 2]}
```

---

I get `True` here:
```
>>> mol = ps.code_to_mol('AdP', rtf)
>>> # This will pass verification in 0.2.3
>>> mol.check_parameters(prm)
False
```

I tried to see which version of `pygen_structures` I installed by printing `ps.__version__` (I installed `0.2.3` of course, as hinted by the comment) but `AttributeError: module 'pygen_structures' has no attribute '__version__'`; I think it's worth adding this attribute to the package.
