### 2020-03-28 a7fc4b3
Warning for segid > 4 chars, fixed error in docs
