### 2020-03-28 552d9d6
Updated docstring for Molecule to add segment length
