### 2020-03-28 #2157 @thesketh
Thanks for your comments, @RMeli. I've added a warning for segids greater than four characters on the Molecule class now.

Well spotted on the wrong type in the doccs - I fixed it in the example before but missed it in that one!

The version string was in `pygen_structures.version.version` before, but I've added it to `pygen_structures.__version__` as well.
