### 2020-04-03 9150522
Version bump to 0.2.4 for PyPI release, update to paper.bib short titles.
