+++ paper.bib
@@ -40,6 +40,7 @@
   author = {Huang, Jing and MacKerell, Alexander D.},
   date = {2013-09},
   journaltitle = {Journal of Computational Chemistry},
+  shortjournal = {J. Comput. Chem.},
   volume = {34},
   pages = {2135--2145},
   issn = {01928651},
@@ -96,8 +97,8 @@
   shorttitle = {{{OpenMM}} 7},
   author = {Eastman, Peter and Swails, Jason and Chodera, John D. and McGibbon, Robert T. and Zhao, Yutong and Beauchamp, Kyle A. and Wang, Lee-Ping and Simmonett, Andrew C. and Harrigan, Matthew P. and Stern, Chaya D. and Wiewiora, Rafal P. and Brooks, Bernard R. and Pande, Vijay S.},
   date = {2017-07-26},
-  journaltitle = {PLOS Computational Biology},
-  shortjournal = {PLOS Computational Biology},
+  journaltitle = {PLoS Computational Biology},
+  shortjournal = {PLoS Comp. Bio.},
   volume = {13},
   pages = {e1005659},
   publisher = {{Public Library of Science}},
@@ -148,6 +149,7 @@
   author = {Riniker, Sereina and Landrum, Gregory A.},
   date = {2015-12},
   journaltitle = {Journal of Chemical Information and Modeling},
+  shortjournal = {J. Chem. Inf. Model.},
   volume = {55},
   pages = {2562--2574},
   publisher = {{American Chemical Society}},
@@ -159,10 +161,11 @@
 }
 
 @article{vmd,
-  title = {{{VMD}} - Visual Molecular Dynamics},
+  title = {{{VMD}} - {{Visual Molecular Dynamics}}},
   author = {Humphrey, William and Dalke, Andrew and Schulten, Klaus},
   date = {1996},
   journaltitle = {Journal of Molecular Graphics},
+  shortjournal = {J. Mol. Graph.},
   volume = {14},
   pages = {33--38},
   doi = {10.1016/0263-7855(96)00018-5}
