### 2020-04-03 917612c
Updates to paper to reflect version changes, PSF format [skip ci]
