+++ paper.md
@@ -43,30 +43,30 @@ KEYWDS    NULL
 EXPDTA    NULL                                                        
 AUTHOR    Marvin                                                      
 REVDAT   1   25-FEB-20         0                                  
-HETATM    1  C   UNK     0       1.322   4.682   0.738  0.00  0.00       \
-...    C+0
-HETATM    2  C   UNK     0      -0.092   4.630   0.732  0.00  0.00       \
-...    C+0
-HETATM    3  C   UNK     0      -0.852   5.626   1.374  0.00  0.00       \
-...    C+0
-HETATM    4  C   UNK     0      -0.210   6.682   2.043  0.00  0.00       \
-...    C+0
-HETATM    5  C   UNK     0       1.194   6.742   2.065  0.00  0.00       \
-...    C+0
-HETATM    6  C   UNK     0       1.956   5.751   1.415  0.00  0.00       \
-...    C+0
-HETATM    7  C   UNK     0       2.119   3.650   0.016  0.00  0.00       \
-...    C+0
-HETATM    8  C   UNK     0       2.390   4.043  -1.467  0.00  0.00       \
-...    C+0
-HETATM    9  C   UNK     0       3.204   3.058  -2.216  0.00  0.00       \
-...    C+0
-HETATM   10  N   UNK     0       1.121   4.261  -2.184  0.00  0.00       \
-...    N+1
-HETATM   11  O   UNK     0       3.531   3.283  -3.404  0.00  0.00       \
-...    O+0
-HETATM   12  O   UNK     0       3.645   1.907  -1.656  0.00  0.00       \
-...    O-1
+HETATM    1  C   UNK     0       1.322   4.682   0.738  0.00  0.00     \
+...      C+0
+HETATM    2  C   UNK     0      -0.092   4.630   0.732  0.00  0.00     \
+...      C+0
+HETATM    3  C   UNK     0      -0.852   5.626   1.374  0.00  0.00     \
+...      C+0
+HETATM    4  C   UNK     0      -0.210   6.682   2.043  0.00  0.00     \
+...      C+0
+HETATM    5  C   UNK     0       1.194   6.742   2.065  0.00  0.00     \
+...      C+0
+HETATM    6  C   UNK     0       1.956   5.751   1.415  0.00  0.00     \
+...      C+0
+HETATM    7  C   UNK     0       2.119   3.650   0.016  0.00  0.00     \
+...      C+0
+HETATM    8  C   UNK     0       2.390   4.043  -1.467  0.00  0.00     \
+...      C+0
+HETATM    9  C   UNK     0       3.204   3.058  -2.216  0.00  0.00     \
+...      C+0
+HETATM   10  N   UNK     0       1.121   4.261  -2.184  0.00  0.00     \
+...      N+1
+HETATM   11  O   UNK     0       3.531   3.283  -3.404  0.00  0.00     \
+...      O+0
+HETATM   12  O   UNK     0       3.645   1.907  -1.656  0.00  0.00     \
+...      O-1
 MASTER        0    0    0    0    0    0    0    0   12    0    0    0
 END
 ```
@@ -80,30 +80,30 @@ Note the lack of annotation in the PDB atom names (labelled as their element sym
 >>> AllChem.EmbedMolecule(mol, AllChem.ETKDGv2())
 0
 >>> print(Chem.MolToPDBBlock(mol))
-ATOM      1  N   PHE A   1      -2.664  -1.853   0.236  1.00  0.00       \
-...    N  
[…]
