### 2020-04-03 #2157 @thesketh
@richardjgowers I've released another version (v0.2.4) on PyPI to reflect @RMeli's recommended changes. I've updated the examples in the article to reflect the new version, could you update the version with whedon? I believe it's an editor-only command
