### 2020-04-13 #2157 @arfon
@thesketh - I made a minor whitespace formatting fix in https://github.com/thesketh/pygen-structures/pull/9 that it would be good to have merged before we accept and publish here.
