### 2020-04-13 #2157 @thesketh
Thanks for the PR, @arfon. The v0.2.4 release already contains all the changes, the Zenodo DOI is 10.5281/zenodo.3739128
